ApiServer: fixed non primitive Long "snapshotLimit" comparsion

+++ server/src/com/cloud/api/ApiServer.java
@@ -208,7 +208,7 @@ public class ApiServer extends ManagerBase implements HttpRequestHandler, ApiSer
         String strSnapshotLimit = configs.get(Config.ConcurrentSnapshotsThresholdPerHost.key());
         if (strSnapshotLimit != null) {
             Long snapshotLimit = NumbersUtil.parseLong(strSnapshotLimit, 1L);
-            if (snapshotLimit <= 0) {
+            if (snapshotLimit.longValue() <= 0) {
                 s_logger.debug("Global config parameter " + Config.ConcurrentSnapshotsThresholdPerHost.toString()
                         + " is less or equal 0; defaulting to unlimited");
             } else {
